+++ open-adexchange-api/open-adexchange-ssp-api/src/main/java/top/openadexchange/openapi/ssp/infra/repository/CacheMetadataRepository.java
@@ -28,7 +28,7 @@ public class CacheMetadataRepository implements MetadataRepository {
     }
 
     @Override
-    public Site getSite(Long siteId) {
+    public Site getSite(Integer siteId) {
         return metadataCacheService.getSite(siteId);
     }
